llama precision experiments

+++ src/fact_completion_scripts/compare_models.py
@@ -89,7 +89,9 @@ def get_model_and_tokenizer(model_name):
             tokenizer_path
         ), transformers.LLaMAForCausalLM.from_pretrained(
             model_name,
-            device_map="auto",  # , load_in_8bit=True, torch_dtype=torch.float16
+            device_map="auto",
+            torch_dtype=torch.float16,
+            # , load_in_8bit=True, torch_dtype=torch.float16
         )
